Implemented script to continuously track devices on the wifi
Implemented script (start_service.py). This script will perform an nmap
scan of the local subnet. It will then view the arp table and find
information about hardware addresses of all found devices. If a device
appears in the arp scan and not the nmap scan, it will ping the device
to determine its status.